Fix breaking change rename of CategoriesHighlights
