--- react/CategoriesHighlighted.js
@@ -1,3 +1,0 @@
-import CategoriesHighlighted from './components/CategoriesHighlighted'
-
-export default CategoriesHighlighted
