+++ react/CategoriesHighlights.js
@@ -1,0 +1,3 @@
+import CategoriesHighlights from './components/CategoriesHighlights'
+
+export default CategoriesHighlights
