+++ react/components/CategoriesHighlights/README.md
