+++ react/components/CategoriesHighlights/constants.js
