+++ react/components/CategoriesHighlights/global.css
